Сделал отдельную логику для разных модальных окон
